Refactor graph processor
