+++ server/src/http_objects.rs
@@ -1061,8 +1061,8 @@ impl From<data_model::Allocation> for Allocation {
             namespace: allocation.namespace,
             compute_graph: allocation.compute_graph,
             compute_fn: allocation.compute_fn,
-            executor_id: allocation.executor_id.to_string(),
-            function_executor_id: allocation.function_executor_id.get().to_string(),
+            executor_id: allocation.target.executor_id.to_string(),
+            function_executor_id: allocation.target.function_executor_id.get().to_string(),
             task_id: allocation.task_id.to_string(),
             invocation_id: allocation.invocation_id.to_string(),
             created_at: allocation.created_at,
