+++ server/state_store/src/lib.rs
@@ -341,7 +341,7 @@ impl IndexifyState {
                                 invocation_id: allocation.invocation_id.clone(),
                                 fn_name: allocation.compute_fn.clone(),
                                 task_id: allocation.task_id.get().to_string(),
-                                executor_id: allocation.executor_id.get().to_string(),
+                                executor_id: allocation.target.executor_id.get().to_string(),
                                 allocation_id: allocation.id.to_string(),
                             },
                         ));
